requirements.txt
Flask
pdfplumber
python-docx
Werkzeug
google-generativeai
fpdf